Remove legacy future imports and share Munin code
